Starting UI

+++ src/IUpdateable.kt
@@ -2,5 +2,5 @@
  * Created by nilot on 14/03/2016.
  */
 interface IUpdateable {
-    fun update(interval:Int)
+    fun update(interval: Double)
 }
